pyqtgraph-zmq/requirements.txt
PyQt5
pyqtgraph
pyzmq